make the cast to float explicit

+++ ubreco/PandoraEventBuildingFlashID/FlashNeutrinoId_tool.cc
@@ -606,12 +606,10 @@ FlashNeutrinoId::SliceCandidate::SliceCandidate()
       m_n_michel_hits_plane0(-std::numeric_limits<int>::max()),
       m_n_michel_hits_plane1(-std::numeric_limits<int>::max()),
       m_n_michel_hits_plane2(-std::numeric_limits<int>::max()),
-    // c14 warning:
-    // implicit conversion from 'std::numeric_limits<int>::type' (aka 'int') to 'float'
-    // changes value from -2147483647 to -2147483648 
-      m_min_lin_braggalgonly_plane0(-std::numeric_limits<int>::max()),
-      m_min_lin_braggalgonly_plane1(-std::numeric_limits<int>::max()),
-      m_min_lin_braggalgonly_plane2(-std::numeric_limits<int>::max()),
+      // deal with c14 warning about implicit int to float
+      m_min_lin_braggalgonly_plane0(-(float)(std::numeric_limits<int>::max())),
+      m_min_lin_braggalgonly_plane1(-(float)(std::numeric_limits<int>::max())),
+      m_min_lin_braggalgonly_plane2(-(float)(std::numeric_limits<int>::max())),
       m_vtx_in_FV(false)
 {
 }
@@ -670,9 +668,10 @@ FlashNeutrinoId::SliceCandidate::SliceCandidate(const art::Event &event, const S
       m_n_michel_hits_plane0(-std::numeric_limits<int>::max()),
       m_n_michel_hits_plane1(-std::numeric_limits<int>::max()),
       m_n_michel_hits_plane2(-std::numeric_limits<int>::max()),
-      m_min_lin_braggalgonly_plane0(-std::numeric_limits<int>::max()),
-      m_min_lin_braggalgonly_plane1(-std::numeric_limits<int>::max()),
-      m_min_lin_braggalgonly_plane2(-std::numeric_limits<int>::max()),
+      // deal with c14 warning about implicit int to float
+      m_min_lin_braggalgonly_plane0(-(float)(std::numeric_limits<int>::max())),
+      m_min_lin_braggalgonly_plane1(-(float)(std::numeric_limits<int>::max())),
+      m_min_lin_braggalgonly_plane2(-(float)(std::numeric_limits<int>::max())),
       m_vtx_in_FV(false)
 {
 }
@@ -738,9 +737,10 @@ FlashNeutrinoId::SliceCandidate::SliceCandidate(const art::Event &event, const S
       m_n_michel_hits_plane0(-std::numeric_limits<int>::max()),
       m_n_michel_hits_plane1(-std::numeric_limits<int>::max()),
       m_n_michel_hits_plane2(-std::numeric_limits<int>::max()),
-      m_min_lin_braggalgonly_plane0(-std::numeric_limits<int>::max()),
-      m_min_lin_braggalgonly_plane1(-std::numeric_limits<int>::max()),
-      m_min_lin_braggalgonly_plane2(-std::numeric_limits<int>::max()),
+      // deal with c14 warning about implicit int to float
+      m_min_lin_braggalgonly_plane0(-(float)(std::numeric_limits<int>::max())),
+      m_min_lin_braggalgonly_plane1(-(float)(std::numeric_limits<int>::max())),
+      m_min_lin_braggalgonly_plane2(-(float)(std::numeric_limits<int>::max())),
       m_vtx_in_FV(false)
 {
     const auto chargeDeposition(this->GetDepositionVector(pfParticleMap, pfParticleToSpacePointMap, spacePointToHitMap, slice));
